# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RPay Payment Gateway API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /merchants/{id}/plan:
    get:
      operationId: getMerchantPlan
      tags: [merchants]
      summary: Get current plan assignment for a merchant (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Merchant plan assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantPlanWithDetails"
        "404":
          description: No plan assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /merchants/{id}/plan/history:
    get:
      operationId: getMerchantPlanHistory
      tags: [merchants]
      summary: Get plan history for a merchant (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanHistory"
...
  /plans/me:
    get:
      operationId: getMyPlan
      tags: [plans]
      summary: Get current merchant's assigned plan
      responses:
        "200":
          description: Merchant plan assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantPlanWithDetails"
        "404":
          description: No plan assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/me/usage:
    get:
      operationId: getMyPlanUsage
      tags: [plans]
      summary: Get current merchant's plan usage counts
      responses:
        "200":
          description: Plan usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanUsage"
        "404":
          description: No plan assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /plans/history:
    get:
      operationId: listPlanHistory
      tags: [plans]
      summary: List all plan history (admin only)
      parameters:
        - name: merchantId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Plan history list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanHistoryListResponse"
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List all plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create a plan (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    put:
      operationId: updatePlan
      tags: [plans]
      summary: Update a plan (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete a plan (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Plan:
      type: object
      required: [id, name, price, monthlyFee, yearlyFee, setupFee, pricing, features, customFeatures, dynamicQrLimit, staticQrLimit, virtualAccountLimit, paymentLinkLimit, payoutLimit, dailyTransactionLimit, monthlyTransactionLimit, settlementFee, depositFee, apiAccess, webhookAccess, providerAccess, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        monthlyFee:
          type: string
        yearlyFee:
          type: string
        setupFee:
          type: string
        pricing:
          type: string
        features:
          type: string
        customFeatures:
          type: string
        dynamicQrLimit:
          type: integer
        staticQrLimit:
          type: integer
        virtualAccountLimit:
          type: integer
        paymentLinkLimit:
          type: integer
        payoutLimit:
          type: integer
        dailyTransactionLimit:
          type: integer
        monthlyTransactionLimit:
          type: integer
        settlementFee:
          type: string
        depositFee:
          type: string
        apiAccess:
          type: boolean
        webhookAccess:
          type: boolean
        providerAccess:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    MerchantPlanWithDetails:
      type: object
      required: [id, merchantId, planId, planName, price, monthlyFee, yearlyFee, setupFee, pricing, features, customFeatures, dynamicQrLimit, staticQrLimit, virtualAccountLimit, paymentLinkLimit, payoutLimit, dailyTransactionLimit, monthlyTransactionLimit, settlementFee, depositFee, apiAccess, webhookAccess, providerAccess, assignedAt, isExpired, status]
      properties:
        id:
          type: integer
        merchantId:
          type: integer
        planId:
          type: integer
        planName:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        monthlyFee:
          type: string
        yearlyFee:
          type: string
        setupFee:
          type: string
        pricing:
          type: string
        features:
          type: string
        customFeatures:
          type: string
        dynamicQrLimit:
          type: integer
        staticQrLimit:
          type: integer
        virtualAccountLimit:
          type: integer
        paymentLinkLimit:
          type: integer
        payoutLimit:
          type: integer
        dailyTransactionLimit:
          type: integer
        monthlyTransactionLimit:
          type: integer
        settlementFee:
          type: string
        depositFee:
          type: string
        apiAccess:
          type: boolean
        webhookAccess:
          type: boolean
        providerAccess:
          type: boolean
        status:
          type: string
          enum: [active, suspended, expired]
        assignedAt:
          type: string
        expiresAt:
          type: ["string", "null"]
        isExpired:
          type: boolean
        daysUntilExpiry:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    PlanInput:
      type: object
      required: [name, pricing, features]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        monthlyFee:
          type: string
        yearlyFee:
          type: string
        setupFee:
          type: string
        pricing:
          type: string
        features:
          type: string
        customFeatures:
          type: string
        dynamicQrLimit:
          type: integer
        staticQrLimit:
          type: integer
        virtualAccountLimit:
          type: integer
        paymentLinkLimit:
          type: integer
        payoutLimit:
          type: integer
        dailyTransactionLimit:
          type: integer
        monthlyTransactionLimit:
          type: integer
        settlementFee:
          type: string
        depositFee:
          type: string
        apiAccess:
          type: boolean
        webhookAccess:
          type: boolean
        providerAccess:
          type: boolean
        isActive:
          type: boolean
    PlanUsage:
      type: object
      required: [dynamicQr, staticQr, virtualAccount, paymentLink, payout, dailyTransaction, monthlyTransaction]
      properties:
        planName:
          type: string
        isExpired:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        daysUntilExpiry:
          type: ["integer", "null"]
        apiAccess:
          type: boolean
        webhookAccess:
          type: boolean
        settlementFee:
          type: string
        depositFee:
          type: string
        dynamicQr:
          type: object
          required: [used, limit]
          properties:
            used: { type: integer }
            limit: { type: integer }
        staticQr:
          type: object
          required: [used, limit]
          properties:
            used: { type: integer }
            limit: { type: integer }
        virtualAccount:
          type: object
          required: [used, limit]
          properties:
            used: { type: integer }
            limit: { type: integer }
        paymentLink:
          type: object
          required: [used, limit]
          properties:
            used: { type: integer }
            limit: { type: integer }
        payout:
          type: object
          required: [used, limit]
          properties:
            used: { type: integer }
            limit: { type: integer }
        dailyTransaction:
          type: object
          required: [used, limit]
          properties:
            used: { type: integer }
            limit: { type: integer }
        monthlyTransaction:
          type: object
          required: [used, limit]
          properties:
            used: { type: integer }
            limit: { type: integer }
    PlanHistory:
      type: object
      required: [id, merchantId, action, createdAt]
      properties:
        id:
          type: integer
        merchantId:
          type: integer
        fromPlanId:
          type: ["integer", "null"]
        toPlanId:
          type: ["integer", "null"]
        toPlanName:
          type: ["string", "null"]
        action:
          type: string
        assignedBy:
          type: ["integer", "null"]
        adminEmail:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        businessName:
          type: ["string", "null"]
        createdAt:
          type: string
    PlanHistoryListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PlanHistory"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...