# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RPay Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  monthlyFee: string;
  yearlyFee: string;
  setupFee: string;
  pricing: string;
  features: string;
  customFeatures: string;
  dynamicQrLimit: number;
  staticQrLimit: number;
  virtualAccountLimit: number;
  paymentLinkLimit: number;
  payoutLimit: number;
  dailyTransactionLimit: number;
  monthlyTransactionLimit: number;
  settlementFee: string;
  depositFee: string;
  apiAccess: boolean;
  webhookAccess: boolean;
  providerAccess: boolean;
  isActive: boolean;
  createdAt: string;
  updatedAt?: string;
}